Moved Currency to Core Module
