+++ src/AdminAssistant.Blazor/Server/WebAPI/v1/CoreModule/CurrencyController.CurrencyResponseDto.cs
@@ -1,4 +1,4 @@
-using AdminAssistant.DomainModel.Modules.AccountsModule;
+using AdminAssistant.DomainModel.Modules.CoreModule;
 using AdminAssistant.Framework.TypeMapping;
 
 namespace AdminAssistant.WebAPI.v1.CoreModule
