+++ src/AdminAssistant.Infra/DAL/EntityFramework/Model/Core/CurrencyEntity.cs
@@ -1,4 +1,4 @@
-using AdminAssistant.DomainModel.Modules.AccountsModule;
+using AdminAssistant.DomainModel.Modules.CoreModule;
 using AdminAssistant.Framework.TypeMapping;
 
 namespace AdminAssistant.Infra.DAL.EntityFramework.Model.Core
