+++ src/AdminAssistant.Test/AutoMapper_WebAPIClientMappingProfile_UnitTest.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Diagnostics.CodeAnalysis;
 using AdminAssistant.DomainModel.Modules.AccountsModule;
+using AdminAssistant.DomainModel.Modules.CoreModule;
 using AutoMapper;
 using FluentAssertions;
 using Xunit;
