+++ src/AdminAssistant.Test/DomainModel/Modules/CoreModule/CQRS/CurrenciesQuery_UnitTest.cs
@@ -1,9 +1,7 @@
 #pragma warning disable CA1707 // Identifiers should not contain underscores
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using AdminAssistant.Infra.DAL.Modules.AccountsModule;
-using AdminAssistant.DomainModel.Modules.AccountsModule;
-using AdminAssistant.DomainModel.Modules.AccountsModule.CQRS;
+using AdminAssistant.Infra.DAL.Modules.CoreModule;
 using Ardalis.Result;
 using FluentAssertions;
 using MediatR;
@@ -11,7 +9,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Moq;
 using Xunit;
 
-namespace AdminAssistant.DomainModel.Modules.Core.CQRS
+namespace AdminAssistant.DomainModel.Modules.CoreModule.CQRS
 {
     public class CurrenciesQuery_Should
     {
