+++ src/AdminAssistant.Test/WebAPI/v1/CoreModule/CurrencyController_UnitTest.cs
@@ -4,8 +4,8 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using AdminAssistant.DomainModel;
-using AdminAssistant.DomainModel.Modules.AccountsModule;
-using AdminAssistant.DomainModel.Modules.AccountsModule.CQRS;
+using AdminAssistant.DomainModel.Modules.CoreModule;
+using AdminAssistant.DomainModel.Modules.CoreModule.CQRS;
 using Ardalis.Result;
 using AutoMapper;
 using FluentAssertions;
