+++ src/AdminAssistant/DomainModel/Factory.cs
@@ -1,5 +1,6 @@
 using AdminAssistant.DomainModel.Modules.AccountsModule.Builders;
 using AdminAssistant.DomainModel.Modules.BudgetModule.Builders;
+using AdminAssistant.DomainModel.Modules.CoreModule.Builders;
 using AdminAssistant.DomainModel.Modules.DocumentsModule.Builders;
 using AdminAssistant.DomainModel.Modules.TasksModule.Builders;
 
@@ -7,21 +8,23 @@ namespace AdminAssistant.DomainModel
 {
     public static class Factory
     {
-        // AccountsModule ...
+        // Accounts Module ...
         public static IBankBuilder Bank => new BankBuilder();
         public static IBankAccountBuilder BankAccount => new BankAccountBuilder();
         public static IBankAccountInfoBuilder BankAccountInfo => new BankAccountInfoBuilder();
         public static IBankAccountTypeBuilder BankAccountType => new BankAccountTypeBuilder();
         public static IBankAccountTransactionBuilder BankAccountTransaction => new BankAccountTransactionBuilder();
-        public static ICurrencyBuilder Currency => new CurrencyBuilder();
 
-        // BudgetModule ...
+        // Budget Module ...
         public static IBudgetBuilder Budget => new BudgetBuilder();
 
-        // DocumentsModule ...
+        // Core Module ...
+        public static ICurrencyBuilder Currency => new CurrencyBuilder();
+
+        // Documents Module ...
         public static IDocumentBuilder Document => new DocumentBuilder();
 
-        // TasksModule ...
+        // Tasks Module ...
         public static ITaskListBuilder TaskList => new TaskListBuilder();
     }
 }
