+++ src/AdminAssistant/DomainModel/Modules/CoreModule/Validation/CurrencyValidator.cs
@@ -1,6 +1,6 @@
 using FluentValidation;
 
-namespace AdminAssistant.DomainModel.Modules.AccountsModule.Validation
+namespace AdminAssistant.DomainModel.Modules.CoreModule.Validation
 {
     public interface ICurrencyValidator : IValidator<Currency>
     {
