+++ src/AdminAssistant/Infra/DAL/Modules/AccountsModule/ICurrencyRepository.cs
@@ -1,6 +1,6 @@
-using AdminAssistant.DomainModel.Modules.AccountsModule;
+using AdminAssistant.DomainModel.Modules.CoreModule;
 
-namespace AdminAssistant.Infra.DAL.Modules.AccountsModule
+namespace AdminAssistant.Infra.DAL.Modules.CoreModule
 {
     public interface ICurrencyRepository : IReadOnlyRepository<Currency>
     {
