+++ src/AdminAssistant/UI/Modules/AccountsModule/IBankAccountEditDialogViewModel.cs
@@ -1,6 +1,7 @@
 using System;
 using System.ComponentModel;
 using AdminAssistant.DomainModel.Modules.AccountsModule;
+using AdminAssistant.DomainModel.Modules.CoreModule;
 using Microsoft.Toolkit.Mvvm.Input;
 using Microsoft.Toolkit.Mvvm.Messaging;
