+++ src/AdminAssistant/UI/Modules/CoreModule/ICoreService.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using AdminAssistant.DomainModel.Modules.AccountsModule;
+using AdminAssistant.DomainModel.Modules.CoreModule;
 
 namespace AdminAssistant.UI.Modules.CoreModule
 {
